refactor separate utils-sketch/{.,layers} + utils
