+++ src/utils-sketch/showCursorPopupBrowserWindow.js
